requirements-test.txt
pytest
mock
requests-mock
pytest-cov